% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_catppuccin_point}
\alias{geom_catppuccin_point}
\alias{geom_catppuccin_col}
\alias{geom_catppuccin_line}
\alias{geom_catppuccin_bar}
\alias{geom_catppuccin_boxplot}
\alias{geom_catppuccin_violin}
\title{Catppuccin Geom Helpers}
\usage{
geom_catppuccin_point(
  mapping = NULL,
  data = NULL,
  flavor = getOption("ggcatppuccin.flavor", "mocha"),
  color = NULL,
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_catppuccin_col(
  mapping = NULL,
  data = NULL,
  flavor = getOption("ggcatppuccin.flavor", "mocha"),
  fill = NULL,
  color = NA,
  ...,
  position = "stack",
  width = NULL,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_catppuccin_line(
  mapping = NULL,
  data = NULL,
  flavor = getOption("ggcatppuccin.flavor", "mocha"),
  color = NULL,
  linewidth = NULL,
  linetype = 1,
  ...,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_catppuccin_bar(
  mapping = NULL,
  data = NULL,
  flavor = getOption("ggcatppuccin.flavor", "mocha"),
  fill = NULL,
  color = NA,
  ...,
  position = "stack",
  width = NULL,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_catppuccin_boxplot(
  mapping = NULL,
  data = NULL,
  flavor = getOption("ggcatppuccin.flavor", "mocha"),
  fill = NULL,
  color = NULL,
  ...,
  outlier.colour = NULL,
  outlier.fill = NULL,
  outlier.shape = 19,
  outlier.size = 1.5,
  outlier.stroke = 0.5,
  outlier.alpha = NULL,
  notch = FALSE,
  notchwidth = 0.5,
  staplewidth = 0,
  varwidth = FALSE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_catppuccin_violin(
  mapping = NULL,
  data = NULL,
  flavor = getOption("ggcatppuccin.flavor", "mocha"),
  fill = NULL,
  color = NULL,
  ...,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, position, na.rm, show.legend, inherit.aes, ...}{Standard ggplot2
geom parameters. See the corresponding ggplot2 geom documentation for details.}

\item{flavor}{Which Catppuccin flavor to use ("mocha", "latte", "frappe", "macchiato").
Defaults to the value set by \code{set_catppuccin_flavor()}, or "mocha" if not set.}

\item{color, colour}{Default point/line/border color; falls back to a Catppuccin
accent if \code{NULL}.}

\item{fill}{Default fill color; falls back to a Catppuccin accent if \code{NULL}.}

\item{width}{Bar/column width.}

\item{orientation}{Orientation of the geom ("x" or "y").}

\item{linewidth, linetype}{Line width and type for line geoms.}

\item{outlier.colour, outlier.fill, outlier.shape, outlier.size, outlier.stroke, outlier.alpha}{Aesthetics for outliers in boxplots.}

\item{notch, notchwidth, staplewidth, varwidth}{Boxplot parameters.}

\item{trim, scale}{Violin plot parameters.}
}
\value{
A ggplot2 layer object.
}
\description{
Convenience wrappers around standard ggplot2 geoms that default to Catppuccin
colors so you can get a styled plot without manually setting aesthetics each time.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(mpg, wt, color = factor(cyl))) +
  geom_catppuccin_point() +
  scale_color_catppuccin()

ggplot(mtcars, aes(factor(cyl), mpg)) +
  geom_catppuccin_col()
}
